requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
django-cors-headers>=3.13
python-dotenv>=1.0
mysqlclient>=2.2.7
bcrypt>=4.3.0
Pillow>=10.0.0